First cut at the client parts.
Import the pgextwlist code base and add archive and platform bits to
prepare hooking into ProcessUtility(). Determining the current binary
platform with no fork allowed was... more fun than anticipated.

It might be a good idea to revise the platform detection code in the
repository server to use the same file based heuristic, plus `uname`
that at least we have as a syscall when doing C code.