Implement client/server protocol in the C parts.
We use libcurl to talk HTTP to the pginstall server, and we parse the
results using the jansson library.

We could have been using the internal PostgreSQL API to parse the JSON
text, but that lib wasn't included in 9.1 and changed in a non trivial
way between 9.2 and 9.3. This way we still have some hope to end up with
a 9.1 onward compatibility.

Only supported just now is the pginstall_available_extensions view and
SRF which makes visible in SQL information from the repository server.