Clean up source code organisation.
Some parts where still unclear about running in the context of the
repository server, where we are using the database server, or the
buildfarm animal where we are using the repository server HTTP API for
communication.

This new organisation moves things where they should be, and begins the
documentation effort of the solution with src/README.md.