Review the Makefile, add INSTALL documentation.
It appears to be harder than it should to build pginstall, this patch is
an attempt at simplifying the process by reviewing the Makefile, adding
an effective `make clean` command for getting out of a confusing
situation.

Also, provide an INSTALL.md file covering the basics and dependencies
you need to manually care about when you want to build pginstall
yourself.