Differenciate between Buildfarm Animal and local builds.
Allow for building an extension locally when given its fullname, without
even trying to contact a repository manager. That is the beginning of
the implementation of a new mode of operation for pginstall, the lone
developer.

Next steps should allow the client to use local archive files without
ever contacting the repository server either, so that a local developer
can benefit from pginstall without any sort of "infrastucture".

After that, implementing the "upstream" setting where a local repository
serves as a proxy when a local file isn't found will allow for fully
integrated mode of operation with a mix of locally built extensions and
centrally provided archives.