Initial support for AWS S3 storage.
The pginstall.ini setup now has support for an S3 section allowing for
uploading archive files to S3. When uploading an archive to the
pginstall server, if some S3 bucket name has been setup, the extension
archive file is uploaded to S3.

It should be possible to then setup an URL (CloudFront or otherwise)
that would serve the static files using a bare HTTP protocol, thus
possible to setup as a pginstall.repository in the clients.

Note that it was already possible to host the archive files on a static
web server where no running pginstall software is required.